requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
future==0.18.2
html5lib==1.1
Pillow==9.0.1
psycopg2==2.9.3
pycodestyle==2.8.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-decouple==3.6
pytz==2021.3
reportlab==3.6.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
webencodings==0.5.1
xhtml2pdf==0.2.5
